refactor: consolidate bitcoin stream files, drop dead public API
- Remove StreamingNativeParser interface + public StreamBlock method +
  BitcoinBlockVisitor[Func] types. Public streaming surface is now just
  StreamBlockIter / BlockStream (iter.Seq2-based).
- Merge bitcoin_native_stream.go and bitcoin_native_stream_iter.go into
  a single bitcoin_native_stream.go. With the visitor/StreamBlock types
  now package-private, the two files were one unit; split only remained
  because the former had public types.
- Replace the internal visitor interface with a plain emit callback
  (func(*BitcoinTransaction) error) consumed directly by the iter loop.
- Rewrite BenchmarkBitcoinStreamBlock against StreamBlockIter.
- Drop the now-dead StreamBlock section from zcash_large_bench_test.go
  (zcash_stream_bench_test.go already benchmarks the streaming path
  end-to-end).
